requirements.txt
asgiref==3.10.0
Django==5.2.7
django-filter==25.2
django-jazzmin==3.0.1
djangorestframework==3.16.1
dotenv==0.9.9
exceptiongroup==1.3.0
Faker==37.11.0
iniconfig==2.3.0
load-dotenv==0.1.0
Markdown==3.9
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
python-dotenv==1.1.1
sqlparse==0.5.3
tomli==2.3.0
typing_extensions==4.15.0
tzdata==2025.2
